Moved fake datasources for instrumented tests to the data-testing module
