+++ data/src/androidTest/kotlin/fobo66/valiutchik/core/model/datasource/FormattingDataSourceImplTest.kt
@@ -18,7 +18,7 @@ package fobo66.valiutchik.core.model.datasource
 
 import androidx.test.filters.SmallTest
 import com.google.common.truth.Truth.assertThat
-import fobo66.valiutchik.core.fake.FakeBankNameNormalizer
+import dev.fobo66.core.data.testing.fake.FakeBankNameNormalizer
 import org.junit.Test
 import java.util.Locale
